Penesito

+++ tests.sh
@@ -1,6 +1,7 @@
 TESTSPORO=25
 TESTSVECTOR=19
 TESTSTLISTA=15
+TESTSABBPORO=0
 nombreBase="tad"
 extension=".cpp"
 zero="0"
@@ -9,10 +10,11 @@ echo "Autocorrector Cuadernillo 1 PED"
 echo "1. Test TPoro"
 echo "2. Test TVectorPoro"
 echo "3. Test TListaPoro"
-echo "4. Todos"
+echo "4. Test TABBPoro"
+echo "5. Todos"
 echo 'Opcion: '
 read Opcion
-if (( $Opcion == 1 || $Opcion == 4 )); then
+if (( $Opcion == 1 || $Opcion == 5 )); then
 	
 	echo "//////////////////////////////////////////////////////////////////"
 	echo "///////////////////////Pruebas TPoro//////////////////////////////"
@@ -37,7 +39,7 @@ if (( $Opcion == 1 || $Opcion == 4 )); then
 	done
 	rm src/*.cpp
 fi
-if (( $Opcion == 2 || $Opcion == 4 )); then
+if (( $Opcion == 2 || $Opcion == 5 )); then
 
 	echo "//////////////////////////////////////////////////////////////////"
 	echo "//////////////////////Pruebas TVectorPoro/////////////////////////"
@@ -62,7 +64,7 @@ if (( $Opcion == 2 || $Opcion == 4 )); then
 	done
 	rm src/*.cpp
 fi
-if (( $Opcion == 3 || $Opcion == 4 )); then
+if (( $Opcion == 3 || $Opcion == 5 )); then
 	echo "//////////////////////////////////////////////////////////////////"
 	echo "///////////////////////Pruebas TListaPoro/////////////////////////"
 	cp src/tlistaporo/*.cpp src/
@@ -78,6 +80,30 @@ if (( $Opcion == 3 || $Opcion == 4 )); then
 		make > basura.txt
 		./tad > $i.txt
 		diff -b $i.txt salidas/tlistaporo/$nombre.sal
+		mv src/tad.cpp src/$nombre
+		rm ./tad
+		rm ./basura.txt
+		rm $i.txt
+		
+	done
+	rm src/*.cpp
+fi
+if (( $Opcion == 4 || $Opcion == 5 )); then
+	echo "//////////////////////////////////////////////////////////////////"
+	echo "///////////////////////Pruebas TABBPoro///////////////////////////"
+	cp src/tabbporo/*.cpp src/
+	for (( i = 1; i <= $TESTSTABBPORO; i++ )); do
+		
+		if (($i >= 10)); then
+			nombre=$nombreBase$i$extension 
+		else
+			nombre=$nombreBase$zero$i$extension 
+		fi
+		echo $nombre
+		mv src/$nombre src/tad.cpp
+		make > basura.txt
+		./tad > $i.txt
+		diff -b $i.txt salidas/tabbporo/$nombre.sal
 		mv src/tad.cpp src/$nombre
 		rm ./tad
 		rm ./basura.txt
